+++ core-ng/src/main/java/core/framework/module/DBConfig.java
@@ -33,7 +33,7 @@ public class DBConfig extends Config {
         this.name = name;
 
         var database = new DatabaseImpl("db" + (name == null ? "" : "-" + name));
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> database.close());
+        context.shutdownHook.add(ShutdownHook.STAGE_6, timeout -> database.close());
         context.backgroundTask().scheduleWithFixedDelay(database.pool::refresh, Duration.ofMinutes(10));
         context.collector.metrics.add(new PoolMetrics(database.pool));
         context.beanFactory.bind(Database.class, name, database);