+++ core-ng/src/main/java/core/framework/module/LogConfig.java
@@ -30,7 +30,7 @@ public class LogConfig extends Config {
         var appender = new KafkaAppender(new KafkaURI(uri));
         appender(appender);
         context.startupHook.start.add(appender::start);
-        context.shutdownHook.add(ShutdownHook.STAGE_8, appender::stop);
+        context.shutdownHook.add(ShutdownHook.STAGE_7, appender::stop);
         context.collector.metrics.add(appender.producerMetrics);
     }