+++ core-ng/src/test/java/core/framework/internal/async/ExecutorImplTest.java
@@ -28,7 +28,7 @@ class ExecutorImplTest {
     @BeforeEach
     void createExecutorImpl() {
         logManager = new LogManager();
-        executor = new ExecutorImpl(1, "test", logManager);
+        executor = new ExecutorImpl(1, "test", logManager, 30_000_000_000L);
     }
 
     @AfterEach